[plugin-analyzer] feat: analyze specific plugin bundle
